+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/publishers/PostPublisher.kt
@@ -9,6 +9,7 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
 import com.github.insanusmokrassar.AutoPostTelegramBot.mediumBroadcastCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.base.commands.deletePost
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers.Chooser
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.chooseCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.sendToLogger
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.DeleteMessage
@@ -35,7 +36,7 @@ private typealias ChatIdMessageIdPair = Pair<ChatId, MessageIdentifier>
 class PostPublisher : Publisher {
     @Transient
     val postPublishedChannel = BroadcastChannel<PostIdListPostMessagesTelegramMessages>(
-        mediumBroadcastCapacity
+        chooseCapacity(mediumBroadcastCapacity)
     )
 
     @Transient